+++ bw2io/__init__.py
@@ -35,7 +35,7 @@ __all__ = [
     'UnlinkedData',
 ]
 
-__version__ = (0, 7, "dev")
+__version__ = (0, 7, "dev1")
 
 
 from .package import BW2Package