+++ CHANGES.md
@@ -1,12 +1,13 @@
 # Changelog
 
-## DEV
+## 0.7 (XXX)
 
 * Support ecoinvent 3.5
 * Drop unused and outdated ecoinvent 31 biosphere flows
 * Add CPC classification to ecoinvent imports
 * Make importers play better with the Activity Browser
 * Add tests for Ecospold2 extraction and importing
+* Add `utils.standardize_method_to_len_3`
 
 ## 0.6 (2018-05-31)